Skip Crust authentication when connecting wallet (#127)
* Fix connect button for Pera by disabling CrustAuth

* Use 0-Algo transaction for Crust authentication

This fixes Crust authentication issues the occur in tools such as Simple Mint and Simple Update.

* Stop asking to authenticate for Crust after auth fails the first time

* Enable Crust auth for all wallets

Crust authentication was only for Pera initially.

* Move `key` attributes to where they are needed

The fixes some annoying errors that appear in the console.

* Disable Crust Auth when connection wallet (again)

Signing a zero-Algo transaction does not work for Crust authentication, and there are issues with signing arbitrary bytes in Pera.